python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "citekit"
version = "0.1.0"
description = "A local SDK that lets AI agents open specific parts of files instead of sending entire documents."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "CiteKit Team" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "sdk", "llm", "pdf", "video", "audio", "rag", "mcp"]
urls = { Repository = "https://github.com/citekit/citekit" }

dependencies = [
    "pydantic>=2.0",
    "pymupdf>=1.24",
    "Pillow>=10.0",
    "google-genai>=1.0",
    "mcp>=1.0",
    "click>=8.0",
]

[project.scripts]
citekit = "citekit.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["citekit"]
